rec_controller/requirements.txt
pycomm3>=1.2.0
pyyaml>=6.0
numpy>=1.20.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0